requirements.txt
lxml>=4.9.0
pandas>=1.3.0
requests>=2.26.0
beautifulsoup4>=4.10.0
python-dotenv>=0.19.0
PyMuPDF>=1.19.0
pdfminer.six>=20211012
